Refactor Hebrew researcher
