+++ packages/yoastseo/src/languageProcessing/languages/he/helpers/createBasicWordForms.js
